rework type of DGNode and DependenceGraph
We'd like to the iterators on node return the right type as well as
the nodes in the graph. That means when graph is LLVMDependenceGraph,
we want to store LLVMDGNode inside, not DGNode<const Value *> etc.
The same with nodes. We want the iterators to return LLVMNode *, not
DGNode<const Value *>.