fix NodesWalk no. 2
This is the same as before, but hopefully for good.
We fixed walk_run_counter for instances of the same
template (e. g. DFS on LLVMNode), but DFS and BFS still
collided (even on same nodes). Now it works and every
NodesWalk on one type of nodes (e. g. LLVMNode) has
it's own counter

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>