points-to: handle inttoptr
this is temprorary workaround - just set it to point
to unknown location. In general, we end up doing this
kind of offen, but we can do better that do it always.
We can keep track of the variables that were casted to int
and back, so we can get more precise analysis. It's to be done...

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>