reaching-defs: don't do strong update with heap-allocated objects
we don't have good mechanism for differentiating heap-allocated objects,
so we represent them as one node. To fix that we would need some
context-sensitivity, but since we don't have it, we just won't do
strong update on heap-allocated objects. This code fails:

  1. struct A *a = alloc_mem(); // calls malloc
  2. a->b = 4;
  3. struct A *b = alloc_mem(); // calls malloc
  4. b->b = 8;   // here RD did strong update, so
  5.             // line 2. was deleted
  6. assert(a->b == 4);

NOTE: with malloc only this is OK, since we would have two
different top-level values. But if it is in subprocedure,
we have only one top-level value (that we cannot tell apart
without the calling context)

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>