tests: rework the framework
We were not sure whether the assertion was called and passed
or wheter we just hit early exit, so slice against test_assert
function that will be linked to the code later, so that we have
control over it. Also with this new approach we can turn few
'let assert fail' tests to normal tests.

There's still few tests that are not converted to the new framework
for different reasons, these will be fixed in future too

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>