RD: add phi node at the beginning of every block
we need that due to mapping back to LLVM nodes.
Mapping is llvm::Value -> RDNode and if the first
node of block is not a definition, that we would need
to map it's llvm::Value to all predecessor's last nodes.
Instead of that we create one new phi node at the beginning
of the block

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>