PSS llvm: reimplement arithmetic insts
if one of the operands is not constant,
check if we have already created one of them.
If not, return unknown memory. This is because
we don't know which of the operands is the one that keeps
pointers and we would need to track both
(and one of the is completely irrelevant)

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>