llvm dg: make BBs that have a self-loop control dependent on itself
When we use the classical algorithm for computing control dependencies,
we never add any self-loop (from design). However, with basic blocks
we need that sometimes, since when a block can jump on itself, then
whether we get to that block again depends on the decision on that block.
Therefore when a block has a self-loop (which appears in optimized code),
make it control dependent on itself.

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>