llvm analyses: do not be so restrictive for funcptr calls
when we have a call of a function via pointer, we always
check whether the prototypes match. But we have been too restrictive.
We can have for example this fragment of LLVM (from phi4 test):

%f.0 = phi void (i32*)*
       [ bitcast (void (i64)* @set1 to void (i32*)*), %%if.then ],
       [ bitcast (void (i64)* @set2 to void (i32*)*), %if.else ]
%1 = ptrtoint i32* %a to i64
%2 = inttoptr i64 %1 to i32*
call void %f.0(i32* %2)

That just breaks the prototypes checks, therefore we need to
be less restrictive in what we allow -- it is an overapproximation
anyway.

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>