PTA: fix FS PTA
The FS PTA sometimes did not reached a fixpoint, because
new information was added when merging information from predecessors,
but in that case the node was not enqueued for new processing.
If we would have the merging implemented in getMemoryObjects,
this would not happen (and it also may save some iterations,
since there would not be any more nodes queued).

fix #137, #136

Signed-off-by: Marek Chalupa <mchqwerty@gmail.com>