do not own any LLVM values
If we create a LLVM value, we put it also into the module.
We create only artificial return or unreachable values,
so this should not break any later analysis.
This way the DG should correspond to the module even
after slicing.
If this turns out to be a problem, we can remove these values anytime
after we are done working with the module.

Signed-off-by: Marek Chalupa <xchalup4@fi.muni.cz>