PTA inv: fix merging pointers from predecessors
When the pointer being merged pointed to a different offset then the invalidated pointer,
we added it to the points-to set again. This led to an infinite loop
since in the next iteration we replace it with invalidated and then add
it again.

Fixes https://github.com/staticafi/symbiotic/issues/88.

Signed-off-by: Marek Chalupa <xchalup4@fi.muni.cz>