Node: refactor and refine dependencies
Add a new dependency: use/user. This is the kind
of data dependency that arises because a node uses
a different node (e.g. as an operand).
We split it off from the data dependency because sometimes
we need to differentiate between this dependency and
"indirect" data dependency that is due to writing and reading
memory.

Signed-off-by: Marek Chalupa <xchalup4@fi.muni.cz>