PTA: re-design global variables
Global variables are now in a special container.
They are processed only once during analysis (at the beginning)
and in the order in which they are created. They still can contain
any type of nodes, e.g. store that may serve for initialization.
It is assumed that the global variables (their points-to sets)
will not change after (the first and the only) processing.