rwg: take allocations as strong definitions where possible
So that the algorithms do not search beyond the definition.
E.g.:

1. ...
2. int a;
3. f(a);

In this case, the algorithm would search also line 1 for definitions
even though there can be none. If we set line 2 as strong definition
of 'a', the algorithm stop the search there. We can do that only if
we know which instance of the object we access, so we do that only
when the address of 'a' is not taken (therefore we must always
access the last instance).