Implement a new style of specifying slicing criteria
This patch adds the -sc option that serves to specify slicing criteria.
This new option allows to specify also file, function and every secondary slicing
criterion may be bound to a specific primary slicing criterion
Further, we now use # as delimiter, so that in future we can
use : in the specification of C++ objects (e.g., A::get).
The -c and -2c options still work as before.

The format for the new criteria is the following:

-sc 'S;S;...'  (i.e., ';' is a delimiter for slicing criteria)
where S ::= C|C  (primary SC | secondary SC)
where C ::= file#function#line#[&][@]object[()]

Where all of function, line, and object can be empty to match _any_ (not all at the
same time). & means that the object is variable, @ means that the object
is global (i.e. global variable or function), () means that the object
is function. If none of &@() is specified, the object may match
any of these. Namely, without () the object will match all instructions
for which we have no points-to information.

The prefix ##... can be left out. That is, instead of '###var',
we can use 'var'.

If only secondary SC is specified (-sc '|foo'), it is evaluated w.r.t
all primary slicing criteria.