tests: Correctly enable assertions even when NDEBUG is set
The problem with the regex replacement in `CMAKE_CXX_FLAGS`is that `NDEBUG`
is usually defined by CMake according to the build type in the
`CMAKE_CXX_FLAGS_TYPE` variable.  Therefore, it was still being defined for
every test.

On the other hand, removing it completely is a bad idea as well since
with `NDEBUG` undefined, some headers may expose symbols that are not and
should not be available in release libraries.

The only place where `assert` is still being used are the fuzzing tests.
This patch explicitly undefines `NDEBUG` before including `cassert` so
that `assert` is not a no-op.  This is fine as POSIX says that:

>  The assert() macro shall be redefined according to the current state of
>  NDEBUG each time <assert.h> is included.

This consequently fixes -Wunused-variable warnings in fuzzing tests.