cmake: mark LLVM's include directories as SYSTEM
.. so that they are passed to the compiler by `-isystem` and not by `-I`.
This makes the approach in `SilenceLLVMWarnings.h` obsolete.

From CMake's documentation:

If the SYSTEM option is given, the compiler will be told the directories are
meant as system include directories on some platforms.  Signalling this
setting  might achieve effects such as the compiler skipping warnings, or these
fixed-install system files not being considered in dependency calculations -
see compiler docs.