build: fix all errors exposed by static LLVM CI
In fact, `LLVMCore` and `LLVMSupport` is always linked as a dependency of
`LLVMIRReader` so linking them twice was redundant.  Most of the command-line
argument errors were caused by `LLVMAnalysis` which was linked first into
`dgllvmdg` and then again into the final executables so the global symbols were
present in two different places causing the errors.